backend/pyproject.toml
[project]
name = "non_counter_parsing"
requires-python = ">=3.9"
version = "0.1.0"
dependencies = [
    "celus-nibbler==12.0.0",
    "fastapi[standard]>=0.115.12",
    "jinja2>=3.1.6",
    "openai-agents>=0.0.6",
    "pandas>=2.2.3",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.0.1",
]

